requirements.txt
fastapi[all]
SQLAlchemy==2.0.48
psycopg2==2.9.11
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]